Default to Windows Server version 1809 for Windows nodes.
Removes all references to 1803, including moving "win1803" directory to
just "windows". A single Windows directory suffices for now, if
necessary in the future we can shard it into directories for each
Windows version.

We've been running tests with Windows 1809 nodes for a couple days in
our fork without major problems:
https://testgrid.k8s.io/google-windows#windows-prototype&width=20.
Testing on Azure is already using 1809:
https://testgrid.k8s.io/sig-windows#Conformance%20acs-engine%20on%20Azure&width=20.